+++ lib/active_record/sort.rb
@@ -3,4 +3,5 @@ require 'arel/extensions'
 
 require File.expand_path(File.join(__FILE__, '../../../ext/active_record/base'))
 
+ActiveRecord::QueryMethods.prepend(ActiveRecord::Sort)
 ActiveRecord::Querying.delegate :sort, to: :all